@@ -4,8 +4,8 @@
   "dependencies": {
     "algoliasearch": "^3.35.1",
     "classnames": "^2.2.6",
-    "dotenv": "^8.1.0",
-    "gatsby": "^2.16.5",
+    "dotenv": "^8.2.0",
+    "gatsby": "^2.17.4",
     "gatsby-image": "^2.2.29",
     "gatsby-plugin-algolia-search": "^0.5.6",
     "gatsby-plugin-disqus": "^1.1.4",
@@ -21,17 +21,17 @@
     "gatsby-remark-external-links": "^0.0.4",
     "gatsby-remark-prismjs": "^3.3.20",
     "gatsby-remark-relative-images": "^0.2.3",
-    "gatsby-transformer-sharp": "^2.2.23",
+    "gatsby-transformer-sharp": "^2.3.0",
     "gatsby-transformer-yaml": "^2.2.15",
-    "node-sass": "^4.12.0",
+    "node-sass": "^4.13.0",
     "normalize.css": "^8.0.1",
     "prismjs": "^1.17.1",
-    "react": "^16.10.2",
+    "react": "^16.11.0",
     "react-async-script-loader": "^0.3.0",
     "react-collapsible": "^2.6.0",
-    "react-dom": "^16.10.2",
+    "react-dom": "^16.11.0",
     "react-helmet": "^5.2.1",
-    "react-instantsearch-dom": "6.0.0-beta.1"
+    "react-instantsearch-dom": "6.0.0-beta.2"
   },
   "devDependencies": {
     "@hot-loader/react-dom": "^16.10.2",
