@@ -1,60 +1,48 @@
-import React, { Component } from 'react'
+import React, { useRef, useEffect } from 'react'
 import Helmet from 'react-helmet'
-import PropTypes from 'prop-types'
+
 import Layout from '../components/Layout'
 import Bird from '../images/twitter_bird.svg'
 
-class News extends Component {
-  /* eslint-disable no-undef */
-  componentDidMount() {
-    if (twttr.widgets) {
-      this.createTimeline()
-    } else {
-      twttr.ready(this.createTimeline)
+export default ({ location }) => {
+  const timeline = useRef(null)
+
+  useEffect(() => {
+    const createTimeline = () => {
+      twttr.widgets.createTimeline(
+        {
+          sourceType: 'profile',
+          screenName: 'catalyst_CX'
+        },
+        timeline.current
+      )
     }
-  }
 
-  createTimeline = () => {
-    twttr.widgets.createTimeline(
-      {
-        sourceType: 'profile',
-        screenName: 'catalyst_CX'
-      },
-      this.timeline
-    )
-  }
-  /* eslint-enable no-undef */
+    if (twttr.widgets) {
+      createTimeline()
+    } else {
+      twttr.ready(createTimeline)
+    }
+  }, [])
 
-  render() {
-    return (
-      <Layout location={this.props.location}>
-        <div className="news">
-          <Helmet title="News" />
-          <section className="container">
-            <h1>
-              What&#39;s <strong>new</strong>?
-            </h1>
-            <div className="news__content">
-              <div className="news__birds">
-                <img src={Bird} alt="bird" width="70" height="70" />
-                <img src={Bird} alt="bird" width="50" height="50" />
-                <img src={Bird} alt="bird" width="40" height="40" />
-              </div>
-              <div
-                className="twitter__timeline"
-                ref={el => {
-                  this.timeline = el
-                }}
-              />
+  return (
+    <Layout location={location}>
+      <div className="news">
+        <Helmet title="News" />
+        <section className="container">
+          <h1>
+            What&#39;s <strong>new</strong>?
+          </h1>
+          <div className="news__content">
+            <div className="news__birds">
+              <img src={Bird} alt="bird" width="70" height="70" />
+              <img src={Bird} alt="bird" width="50" height="50" />
+              <img src={Bird} alt="bird" width="40" height="40" />
             </div>
-          </section>
-        </div>
-      </Layout>
-    )
-  }
+            <div className="twitter__timeline" ref={timeline} />
+          </div>
+        </section>
+      </div>
+    </Layout>
+  )
 }
-News.propTypes = {
-  location: PropTypes.object.isRequired
-}
-
-export default News
