@@ -1,25 +1,21 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React from 'react'
 
-export default class HTML extends React.Component {
-  render() {
-    return (
-      <html {...this.props.htmlAttributes}>
-        <head>
-          <meta charSet="utf-8" />
-          <meta httpEquiv="x-ua-compatible" content="ie=edge" />
-          <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-          {this.props.headComponents}
-
-        </head>
-        <body {...this.props.bodyAttributes}>
-          {this.props.preBodyComponents}
-          <div key={`body`} id="___gatsby" dangerouslySetInnerHTML={{ __html: this.props.body }} />
-          {this.props.postBodyComponents}
-        </body>
-        <script
-          dangerouslySetInnerHTML={{
-            __html: `window.twttr = (function(d, s, id) {
+export default props => (
+  <html {...props.htmlAttributes}>
+    <head>
+      <meta charSet="utf-8" />
+      <meta httpEquiv="x-ua-compatible" content="ie=edge" />
+      <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
+      {props.headComponents}
+    </head>
+    <body {...props.bodyAttributes}>
+      {props.preBodyComponents}
+      <div key={`body`} id="___gatsby" dangerouslySetInnerHTML={{ __html: props.body }} />
+      {props.postBodyComponents}
+    </body>
+    <script
+      dangerouslySetInnerHTML={{
+        __html: `window.twttr = (function(d, s, id) {
                 var js, fjs = d.getElementsByTagName(s)[0],
                   t = window.twttr || {};
                 if (d.getElementById(id)) return t;
@@ -34,21 +30,8 @@ export default class HTML extends React.Component {
                 };
 
                 return t;
-              }(document, "script", "twitter-wjs"));`,
-          }}
-        />
-      </html>
-    );
-  }
-}
-
-HTML.propTypes = {
-  htmlAttributes: PropTypes.object,
-  headComponents: PropTypes.array,
-  bodyAttributes: PropTypes.object,
-  preBodyComponents: PropTypes.array,
-  body: PropTypes.string,
-  postBodyComponents: PropTypes.array,
-};
-
-{/*<link rel="stylesheet" href="https://cdn.jsdelivr.net/docsearch.js/2.5/docsearch.min.css" />*/}
+              }(document, "script", "twitter-wjs"));`
+      }}
+    />
+  </html>
+)
